requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.3
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
ecdsa==0.14.1
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pipenv==2020.11.15
pkg-resources==0.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
python-dotenv==0.17.0
python-jose==3.2.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
virtualenv-clone==0.5.4
whitenoise==5.2.0
